0803
·新增Spitter脚本
·调整连关发枪时机

+++ 2001/ze_hr_dead_center/save_data.js
@@ -4,7 +4,7 @@ import { CSPlayerPawn, Instance, CSGearSlot } from "cs_script/point_script";
  * 数据存读取脚本
  * 此脚本用于实现回合结算后仍能恢复连关数据
  * 此脚本由皮皮猫233编写
- * 2026/7/23
+ * 2026/8/3
  */
 
 const weaponSlots = [CSGearSlot.RIFLE, CSGearSlot.PISTOL];
@@ -53,12 +53,12 @@ Instance.OnScriptInput("ReadData", () => {
         Instance.ServerCommand("c_revive @t");
     });
 
-    // 延迟5秒后读取数据
-    Delay(5, () => {
+    // 延迟2秒后读取数据
+    Delay(2, () => {
         Instance.ServerCommand("say **正在读取连关数据**");
         const players = /** @type {CSPlayerPawn[]} */ (Instance.FindEntitiesByClass("player"));
         for (const player of players) {
-            if (!player || !player.IsValid()) continue;
+            if (!player.IsValid()) continue;
 
             // 检查是否被存储为上局存活的玩家
             if (playerData.has(player)) {
@@ -75,19 +75,22 @@ Instance.OnScriptInput("ReadData", () => {
                     }
                 }
             } else {
-                const playerController = player.GetPlayerController();
-                if (playerController && playerController.IsValid()) {
-                    Instance.ServerCommand("c_infect #" + playerController.GetPlayerSlot());
-                    player.Teleport({ position: { x: 0, y: 0, z: -3560 } });
-                    Delay(0.1, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
-                    Delay(0.2, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
-                    Delay(0.3, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
+                Delay(5, () => {
+                    if (!player.IsValid()) return;
+                    const playerController = player.GetPlayerController();
+                    if (playerController && playerController.IsValid()) {
+                        Instance.ServerCommand("c_infect #" + playerController.GetPlayerSlot());
+                        player.Teleport({ position: { x: 0, y: 0, z: -3560 } });
+                        Delay(0.1, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
+                        Delay(0.2, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
+                        Delay(0.3, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
 
-                    // 延迟1秒后检测是否仍为人类防止部分社区禁止c_infect指令
-                    Delay(1, () => {
-                        if (player.IsValid() && player.GetTeamNumber() === 3) player.Kill();
-                    });
-                }
+                        // 延迟1秒后检测是否仍为人类防止部分社区禁止c_infect指令
+                        Delay(1, () => {
+                            if (player.IsValid() && player.GetTeamNumber() === 3) player.Kill();
+                        });
+                    }
+                });
             }
         }
         playerData.clear();
